candle chart error fix
